cd/requirements.txt
joblib
pandas
scikit-learn
azureml-core
